fix usage of VertexStream and adjust data representation.
Before this change, a `VertexStream` contained a slice of `u8` data
as well as a stride.

The size of the vertex attribute was wrongly assumed to be data.len()
which would lead to assertion violations in the C code.

The structure of `VertexStream` has been changed to contain a poitner to
the buffer data, the size of the vertex attribute the `VertexStream`
describes and a stride.

Additionally, new methods have been added to the `VertexStream` type to
ease its creation.