Add a limits.h header included via -isystem
As it turns out, the upstream code also uses <limits.h>, but it didn't
fail to compile on Windows as it was able to use the host system INT_MAX
value. Two problems with that : it's potentially the wrong value, AND it
doesn't compile on Linux. This patch should fix that by forcing the
inclusion of our custom headers before the system ones, as well as
adding the missing header.