Make build_meshlets call meshopt_optimizeMeshlet automatically
Latest meshoptimizer adds a function that should be called on meshlet
data if it's used for rendering directly. While it is separate in
meshopt's API, it requires access to mutable meshlets and Meshlet struct
in Rust has immutable slices. The cost of just calling this after build
is low so for now we opt to just call it so that users of the library
automatically get optimized meshlets.

Also fix comment for build_meshlets which was not accurate.